refactor: split pdf2md tests into fast (default) and VLM (on-demand)
- New conftest.py: shared fixtures (minimal_pdf, test_pdf, real_pdf_or_skip)
- test_integration.py: fast tests only — CLI I/O, JSON, error handling, all
  using minimal_pdf (never hangs, 60 tests total with unit suite)
- test_vlm.py: real-PDF quality tests (standard + VLM pipeline), run on demand
  with 'uv run pytest tests/test_vlm.py -v'
- pyproject.toml: --ignore=tests/test_vlm.py so default suite stays fast

Previously default suite would hang because TestFullConversion (real PDF,
docling) was in test_integration.py.  Now 'uv run pytest tests/' is always
safe.
